pyproject.toml
[project]
name = "poly"
version = "0.1.0"
description = "polymakret tuul"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "matplotlib>=3.10.8",
    "numpy>=2.4.0",
    "openrouter>=0.1.1",
    "python-dotenv>=1.0.1",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "tqdm>=4.67.1",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["utils"]
